chat_backup/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.54.0
qdrant-client==1.12.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
markdown==3.7
beautifulsoup4==4.12.3
tiktoken==0.8.0
